back: Add getSiblings method in Region Controller.
This commit introduces a new method getSiblings in regionController.js.
This method seeks to return sibling regions for a given region ID and
hierarchy ID. It also modifies the Hierarchy model to include a
getSiblings method and updates the regionRoutes.js to add a new route
for this functionality.

Issue: #159

Signed-off-by: Nikolay Martyanov <ohmspectator@gmail.com>